week-3/upload-parquet-to-gs.requirements.txt
google-cloud-storage==2.7.0
pandas==1.5.3
pyarrow==11.0.0
